#3 refactor: 거래 목록 필터 hook으로 분리
